Prepare for external scanner
To simplify having a custom external scanner, I've moved JS scanner into
`tree-sitter-javascript/scanner.h`. This will allow us to have custom
scanner implementation and still allow us to easily update the upstream
scanner.
